Fix current AmqpClient network reconnection
This fixes the current AmqpClient reconnection so that it no longer has
an exception inside of an exception which was preventing it from
reconnecting. This adds the needed re-connection logic so that if it
attempts to connect and fails, it will continue to attempt to reconnect
until it's able to connect successfully.